added documentation for objective_functions and optimizer class, created two new files to host jax wrapper methods for better readability
